requirements.txt
psutil
matplotlib
numpy
Pillow